router: Move tests to `util/errors` module
These tests don't appear to be related to our routing code at all, so this commit moves them into the module where most of the tested code comes from.
