+++ src/router.rs
@@ -179,63 +179,3 @@ pub fn build_axum_router(state: AppState) -> Router<()> {
         })
         .with_state(state)
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::middleware::log_request::CauseField;
-    use crate::util::errors::{
-        bad_request, cargo_err, forbidden, internal, not_found, AppError, BoxedAppError,
-    };
-    use axum::response::IntoResponse;
-    use diesel::result::Error as DieselError;
-    use http::StatusCode;
-    use serde::de::Error;
-
-    #[test]
-    fn http_error_responses() {
-        // Types for handling common error status codes
-        assert_eq!(bad_request("").response().status(), StatusCode::BAD_REQUEST);
-        assert_eq!(forbidden().response().status(), StatusCode::FORBIDDEN);
-        assert_eq!(
-            BoxedAppError::from(DieselError::NotFound)
-                .response()
-                .status(),
-            StatusCode::NOT_FOUND
-        );
-        assert_eq!(not_found().response().status(), StatusCode::NOT_FOUND);
-
-        // cargo_err errors are returned as 200 so that cargo displays this nicely on the command line
-        assert_eq!(cargo_err("").response().status(), StatusCode::OK);
-
-        // Inner errors are captured for logging when wrapped by a user facing error
-        let response = "-1"
-            .parse::<u8>()
-            .map_err(|err| err.chain(internal("middle error")))
-            .map_err(|err| err.chain(bad_request("outer user facing error")))
-            .unwrap_err()
-            .into_response();
-        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-        assert_eq!(
-            response.extensions().get::<CauseField>().unwrap().0,
-            "middle error caused by invalid digit found in string"
-        );
-
-        // All other error types are converted to internal server errors
-        assert_eq!(
-            internal("").response().status(),
-            StatusCode::INTERNAL_SERVER_ERROR
-        );
-        assert_eq!(
-            BoxedAppError::from(serde_json::Error::custom("ExpectedColon"))
-                .response()
-                .status(),
-            StatusCode::INTERNAL_SERVER_ERROR
-        );
-        assert_eq!(
-            BoxedAppError::from(::std::io::Error::new(::std::io::ErrorKind::Other, ""))
-                .response()
-                .status(),
-            StatusCode::INTERNAL_SERVER_ERROR
-        );
-    }
-}
